Corrección de puntos a valores decimales

+++ backend/apps/users/models.py
@@ -55,7 +55,7 @@ class BusinessConfig(models.Model):
     show_staff_widget = models.BooleanField(default=True)
     show_new_customers_widget = models.BooleanField(default=True)
     show_agenda_widget = models.BooleanField(default=True)
-    points_per_euro = models.IntegerField(default=1)
+    points_per_euro = models.DecimalField(max_digits=5, decimal_places=2, default=1.0)
 
     created_at = models.DateTimeField(auto_now_add=True)
     updated_at = models.DateTimeField(auto_now=True)
